chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/subtree-visibility-053.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Subtree Visibility: accessibility scrollToMakeVisible()</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<meta name="assert" content="subtree-visibility auto subtrees are exposed by accessibility scrollToMakeVisible">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.spacer {
  height: 3000px;
}
.auto {
  subtree-visibility: auto;
}
</style>

<div class=spacer></div>
<div id="hidden" class=auto>
  foo
  <div id="child" tabindex="0">
    bar
  </div>
</div>
<div class=spacer></div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async(t) => {
  const hiddenEl = document.getElementById("hidden");
  let axHidden = axElementById("hidden");
  t.step(() => { assert_equals(axHidden.childrenCount, 3, "Child count after acquire"); });

  axElementById("child").scrollToMakeVisible();

  // If this is frame 1, then during frame 2, we will determine that we're
  // visible and issue notifications of viewport intersection. However, the
  // visibility notification is delayed to frame 3 past the rAF callbacks. So,
  // check the result in frame 4.
  requestAnimationFrame(() => { // Frame 2.
    requestAnimationFrame(() => { // Frame 3.
      requestAnimationFrame(() => { // Frame 4.
        axHidden = axElementById("hidden");
        t.step(() => {
          assert_equals(axHidden.childrenCount, 2, "Child count after activation");
        });
        t.done();
      });
    });
  });
}, "Accessiblility scrollToMakeVisible() causes activatable hidden tree to activate");

</script>
